+++ playground/Elders.Cronus.Sample.Projections/Program.cs
@@ -1,16 +1,11 @@
 ﻿using System;
-using System.Linq;
 using System.Reflection;
 using Elders.Cronus.Pipeline.Config;
 using Elders.Cronus.Pipeline.Hosts;
 using Elders.Cronus.Pipeline.Transport.RabbitMQ.Config;
 using Elders.Cronus.Sample.Collaboration.Users.Commands;
-using Elders.Cronus.Sample.Collaboration.Users.DTOs;
 using Elders.Cronus.Sample.Collaboration.Users.Projections;
-using Elders.Cronus.Sample.CommonFiles;
 using Elders.Cronus.Sample.IdentityAndAccess.Accounts.Commands;
-using NHibernate;
-using NHibernate.Mapping.ByCode;
 using Elders.Cronus.IocContainer;
 using Elders.Cronus.UnitOfWork;
 
@@ -29,6 +24,7 @@ namespace Elders.Cronus.Sample.Projections
             var cfg = new CronusSettings(container)
                 .UseContractsFromAssemblies(new Assembly[] { Assembly.GetAssembly(typeof(RegisterAccount)), Assembly.GetAssembly(typeof(CreateUser)) })
                 .UseProjectionConsumer(consumer => consumer
+                .SetNumberOfConsumerThreads(5)
                     .WithDefaultPublishersWithRabbitMq()
                     .UseRabbitMqTransport()
                     .UseProjections(h => h
@@ -48,21 +44,21 @@ namespace Elders.Cronus.Sample.Projections
             host.Stop();
         }
 
-        static ISessionFactory BuildSessionFactory()
-        {
-            var typesThatShouldBeMapped = Assembly.GetAssembly(typeof(UserProjection)).GetExportedTypes().Where(t => t.Namespace.EndsWith("DTOs"));
-            var cfg = new NHibernate.Cfg.Configuration();
-            Action<ModelMapper> customMappings = modelMapper =>
-            {
-                modelMapper.Class<User>(mapper =>
-                {
-                    mapper.Property(pr => pr.Email, prmap => prmap.Unique(true));
-                });
-            };
+        //static ISessionFactory BuildSessionFactory()
+        //{
+        //    var typesThatShouldBeMapped = Assembly.GetAssembly(typeof(UserProjection)).GetExportedTypes().Where(t => t.Namespace.EndsWith("DTOs"));
+        //    var cfg = new NHibernate.Cfg.Configuration();
+        //    Action<ModelMapper> customMappings = modelMapper =>
+        //    {
+        //        modelMapper.Class<User>(mapper =>
+        //        {
+        //            mapper.Property(pr => pr.Email, prmap => prmap.Unique(true));
+        //        });
+        //    };
 
-            cfg = cfg.AddAutoMappings(typesThatShouldBeMapped, customMappings);
-            cfg.CreateDatabase_AND_OVERWRITE_EXISTING_DATABASE();
-            return cfg.BuildSessionFactory();
-        }
+        //    cfg = cfg.AddAutoMappings(typesThatShouldBeMapped, customMappings);
+        //    cfg.CreateDatabase_AND_OVERWRITE_EXISTING_DATABASE();
+        //    return cfg.BuildSessionFactory();
+        //}
     }
 }