+++ src/Elders.Cronus.Persistence.Cassandra/Config/CassandraEventStoreSettings.cs
@@ -32,6 +32,7 @@ namespace Elders.Cronus.Persistence.Cassandra.Config
         {
             var cluster = Cluster
                 .Builder()
+                .WithRetryPolicy(new EventStoreNoHintedHandOff())
                 .WithConnectionString(connectionString)
                 .Build();
             self.Session = cluster.ConnectAndCreateDefaultKeyspaceIfNotExists();
@@ -57,6 +58,29 @@ namespace Elders.Cronus.Persistence.Cassandra.Config
             var storageManager = new CassandraEventStoreStorageManager(self.Session, self.EventStoreTableNameStrategy);
             storageManager.CreateStorage();
             return self;
+        }
+    }
+
+    public class EventStoreNoHintedHandOff : IRetryPolicy
+    {
+        public RetryDecision OnReadTimeout(IStatement query, ConsistencyLevel cl, int requiredResponses, int receivedResponses, bool dataRetrieved, int nbRetry)
+        {
+            if (nbRetry != 0)
+                return RetryDecision.Rethrow();
+
+            return receivedResponses >= requiredResponses && !dataRetrieved
+                       ? RetryDecision.Retry(cl)
+                       : RetryDecision.Rethrow();
+        }
+
+        public RetryDecision OnUnavailable(IStatement query, ConsistencyLevel cl, int requiredReplica, int aliveReplica, int nbRetry)
+        {
+            return RetryDecision.Rethrow();
+        }
+
+        public RetryDecision OnWriteTimeout(IStatement query, ConsistencyLevel cl, string writeType, int requiredAcks, int receivedAcks, int nbRetry)
+        {
+            return RetryDecision.Rethrow();
         }
     }